+++ src/main/java/wooteco/chess/domain/judge/Judge.java
@@ -9,13 +9,13 @@ public class Judge {
 
     private static final double PAWN_DUPLICATED_DEDUCTION = 0.5;
 
-    public double getScoreByTeam(final Board board, final Team team) {
+    public static double getScoreByTeam(final Board board, final Team team) {
         double defaultScore = board.getDefaultScore(team);
         double deduction = board.countDuplicatedPawns(team) * PAWN_DUPLICATED_DEDUCTION;
         return defaultScore - deduction;
     }
 
-    public Optional<Team> findWinner(final Board board) {
+    public static Optional<Team> findWinner(final Board board) {
         if (board.hasKing(Team.BLACK) && board.hasKing(Team.WHITE)) {
             return Optional.empty();
         }
